fix(tesseract): Split Tesseract to avoid a racy cleanup
Relying on the value behind `Arc::strong_count` to remain unchanged in
the Drop::drop impl is not a good idea. The docs even warn about this:

> Another thread can change the strong count at any time, including
> potentially between calling this method and acting on the result.

So we resolve this by splitting Tesseract's state into a separate
private struct and keeping that in an `Arc` while not implementing
`Clone` on this internal struct, ensuring there is always one instance
of this struct, no matter why many clone of `Tesseract` are created. We
then move the `Drop` impl also on the internal struct. This eliminates
the need to keep track of references ourselves and avoids the
potential race.

As a side-effect, this also removes `Arc` from individual fields, which
reduces the number of atomic refs, that are not free.